+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 5.4.3
-  - Changes from 5.4.1
+  - Changes from 5.4.2
     - Bugfixes
       - #3254 Fixed a bug that could end up hiding roundabout instructions
       - #3260 fixed a bug that provided the wrong location in the arrival instruction